# Conflicts:
#   moamoa_back-end/build.gradle
#   moamoa_back-end/src/main/java/com/ssafy/moamoa/controller/ProjectController.java
#   moamoa_back-end/src/main/java/com/ssafy/moamoa/domain/dto/TechStackForm.java
#   moamoa_back-end/src/main/java/com/ssafy/moamoa/domain/entity/Profile.java
#   moamoa_back-end/src/main/java/com/ssafy/moamoa/repository/ProjectAreaRepository.java
#   moamoa_back-end/src/main/java/com/ssafy/moamoa/repository/querydsl/ProfileRepositoryCustom.java
#   moamoa_back-end/src/main/java/com/ssafy/moamoa/repository/querydsl/ProfileRepositoryImpl.java
#   moamoa_back-end/src/main/java/com/ssafy/moamoa/service/InitService.java
#   moamoa_back-end/src/main/java/com/ssafy/moamoa/service/ProjectService.java
#   moamoa_back-end/src/main/java/com/ssafy/moamoa/service/TechStackService.java
#   moamoa_back-end/src/main/resources/application.yml

+++ moamoa_back-end/src/main/java/com/ssafy/moamoa/controller/ProjectController.java
@@ -20,8 +20,13 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RequestPart;
 import org.springframework.web.multipart.MultipartFile;
 
+import com.ssafy.moamoa.domain.ProjectCategory;
 import com.ssafy.moamoa.domain.dto.ProjectDetail;
 import com.ssafy.moamoa.domain.dto.ProjectForm;
+
+import com.ssafy.moamoa.domain.entity.Project;
+import com.ssafy.moamoa.domain.entity.TechStack;
+
 import com.ssafy.moamoa.service.ProjectService;
 import com.ssafy.moamoa.service.TeamService;
 
@@ -38,12 +43,22 @@ public class ProjectController {
 	private final ProjectService projectService;
 	private final TeamService teamService;
 
-	@ApiOperation(value = "자기가 속한 프로젝트/스터디 조회",
-		notes = "자기가 속한 프로젝트/스터디를 조회한다.")
-	@GetMapping
+	@ApiOperation(value = "자기가 속한 프로젝트 조회",
+		notes = "자기가 속한 프로젝트를 조회한다.")
+	@GetMapping("/project")
 	public ResponseEntity<?> showProjects(Authentication authentication) throws Exception {
 		UserDetails userDetails = (UserDetails)authentication.getPrincipal();
-		List<ProjectForm> projectForms = projectService.findByUser(Long.valueOf(userDetails.getUsername()));
+		List<ProjectForm> projectForms = projectService.findByUser(Long.valueOf(userDetails.getUsername()),
+			ProjectCategory.PROJECT);
+		return new ResponseEntity<List<ProjectForm>>(projectForms, HttpStatus.OK);
+	}
+
+	@ApiOperation(value = "자기가 속한 스터디 조회",
+		notes = "자기가 속한 스터디를 조회한다.")
+	@GetMapping("/study")
+	public ResponseEntity<?> showStudies(Authentication authentication) throws Exception {
+		UserDetails userDetails = (UserDetails)authentication.getPrincipal();
+		List<ProjectForm> projectForms = projectService.findByUser(Long.valueOf(userDetails.getUsername()), ProjectCategory.STUDY);
 		return new ResponseEntity<List<ProjectForm>>(projectForms, HttpStatus.OK);
 	}
 
@@ -51,7 +66,7 @@ public class ProjectController {
 		notes = "팀 페이지 open")
 	@GetMapping("/detail")
 	public ResponseEntity<?> accessProject(@RequestParam("projectId") Long projectId, Authentication authentication) throws Exception {
-		ProjectDetail projectDetail = projectService.accessProject(projectId);
+		ProjectDetail projectDetail = projectService.accessProject(projectId, 1);
 
 		// 로그인 한 상태
 		if(authentication != null)
@@ -71,8 +86,11 @@ public class ProjectController {
 	public ResponseEntity<?> createProject(@RequestPart(value = "projectForm") ProjectForm projectForm, @RequestPart(value = "file") MultipartFile file, Authentication authentication) throws Exception {
 		UserDetails userDetails = (UserDetails)authentication.getPrincipal();
 		projectForm.setUserId(Long.valueOf(userDetails.getUsername()));
-		projectService.creatProject(projectForm,file);
-		return new ResponseEntity<>(HttpStatus.OK);
+
+		ProjectDetail projectDetail = projectService.creatProject(projectForm,file);
+		projectDetail.setLeader(true);
+		return new ResponseEntity<ProjectDetail>(projectDetail, HttpStatus.OK);
+
 	}
 
 	@ApiOperation(value = "프로젝트/스터디 수정",
@@ -87,8 +105,12 @@ public class ProjectController {
 		{
 			throw new Exception("팀장이 아닙니다.");
 		}
-		projectService.updateProject(Long.valueOf(userDetails.getUsername()), projectForm,file);
-		return new ResponseEntity<>(HttpStatus.OK);
+
+		projectForm.setUserId(Long.valueOf(userDetails.getUsername()));
+		ProjectDetail projectDetail = projectService.updateProject(projectForm,file);
+		projectDetail.setLeader(true);
+		return new ResponseEntity<ProjectDetail>(projectDetail, HttpStatus.OK);
+
 	}
 
 	@ApiOperation(value = "프로젝트/스터디 삭제",
@@ -101,7 +123,8 @@ public class ProjectController {
 		{
 			throw new Exception("팀장이 아닙니다.");
 		}
-		projectService.deleteProject(Long.valueOf(userDetails.getUsername()));
+		projectForm.setUserId(Long.valueOf(userDetails.getUsername()));
+		projectService.deleteProject(projectForm);
 		return new ResponseEntity<>(HttpStatus.OK);
 	}
 
@@ -114,6 +137,10 @@ public class ProjectController {
 		if(!teamService.checkLeader(Long.valueOf(userDetails.getUsername()), projectForm.getProjectId()))
 		{
 			throw new Exception("팀장이 아닙니다.");
+		}
+		if(Long.valueOf(userDetails.getUsername()).equals(projectForm.getUserId()))
+		{
+			throw new Exception("팀장은 강퇴할 수 없습니다.");
 		}
 		projectService.deleteMember(projectForm);
 		return new ResponseEntity<>(HttpStatus.OK);
